requirements.txt
hydra-core==1.3.2
imageio==2.21.1
imgaug==0.4.0
matplotlib==3.5.2
mmcv==2.0.0
numpy==1.23.1
omegaconf==2.2.2
opencv_python==4.6.0.66
pycpt==0.0.1
PyYAML==6.0
rasterio==1.3.6
richdem==0.3.4
scikit_image==0.19.3
seaborn==0.12.2
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
